Use version file to maintain orca version
This commit introduces a version file, ORCA_VERSION, that records the
orca version. Previously, this version was maintained in multiple
places, and every time the orca version was bumped, all of these places
had to be changed (see fbc0c5c1fc for an example). This was an
error-prone process.

Now, the orca version only has to be changed in the ORCA_VERSION file.

Authored-by: Chris Hajas <chajas@pivotal.io>