Make GDD tests deterministic
GDD tests framework now acquire the desired lock by updating the nth tuple
in a segment instead of a specified value, so even a hash algorithm changed,
the tests will not be affected. This method works fine except that a segment
has not enough tuples to provide the nth tuple. Fix is simple, enlarge the
test tables from 20 rows to 100 rows.

Authored-by: Ning Yu nyu@pivotal.io