Fix crash when retrieving opclass in query translation to DXL
This crash was occurring when running the partition_indexing and
partition_ddl schedules in a parallel schedule with ORCA. Specifically,
a `select count(*) from part_table` was running simultaneously with an
`\d` on a different table.

We found that the server was crashing when referencing the opclass in
`if (policy->opclasses[i] == default_opclass)`. When looking at the
disassembly via the core dump and tracing the addresses, we saw that the
reference to opclass from rel -> policy -> opclasses was correct
and wasn't getting overwritten; however, the reference pointed to in the
register (which should match this address) was an invalid memory
address.

Storing opclasses in a variable seems to fix the issue without any
change to the logic.