Mark casts from base type to domain as binary-coercible in ORCA
This was causing an ORCA assertion to fail and fall back, as we asserted
that if a type is not binary coercible, then it had a cast function.

Although this code considered domain->base type as binary coercible, it
did not consider base type->domain as binary coercible. However,
coercion paths of COERCION_PATH_RELABELTYPE are binary coercible, so we
should mark them as such for ORCA.