Include DP alternative when number of joins exceeds threshold in Orca's DPv2 transform
Previously, when the number of joins exceeded the join threshold, we
would only take the one lowest cost alternative for groups exceeding
this threshold. However, we cost the greedy, mincard, and query
alternatives as -1 internally to this transform to ensure we always
generate those three plan alternatives. This meant we didn't generate a
DP alternative if numJoins > threshold, which isn't the intention here.

Now, we ensure we use the lowest cost DP group, in addition to
generating the other alternatives.

Additionally, the greedy xform wasn't kept in the heap, so we fix that
too.