Penalize cross joins in Orca's DPv2 xform similarly to real cost model
Previously, the DPv2 transform would penalize cross joins by multiplying
the total cardinality by a constant, causing the total cost of the
entire tree to increase by that constant. For example, if there was a
relatively small cross join, at the bottom of a tree, that returned 100
rows and penalized it by 5X, the cardinality of the above join would
also get multiplied by 5X. This could result in a very small cross
product near the bottom causing the overall cost of the tree to increase
by a much larger amount.

Instead, we accumulate the cost separately and use the same constant
that is used in the cost model during optimization.