Penalize cross products when selecting initial join in Orca's DPv2 greedy alg
While we penalized cross joins for the remaining joins in greedy, we did
not for the first join, which in some cases selected a cross product.
This ended up selecting a poor join order in many cases and went against
the intent of the alternative being generated, which is to minimize
cross products.

Additionally, the greedy xform wasn't kept in the heap, so we include
that now too.