Support partitioned tables with partitions of mixed storage types
While partitioned tables don't have a storage type, Orca treats
partitioned tables as a single unit; and we need to assign it as some
storage type to make certain decisions (eg: not generating an index scan on
an ao table).

If all partitions are of the same type (ao/heap/co), we treat the
partitioned table as that type. However, if it contains mixed partitions
(eg: some heap, some ao), we mark it as storage type "MixedParittioned",
and are more conservative when generating some scans.

We also fall back for now if any partition is external, as we need to handle
distribution differently in this case.